feat(Dao): 조회 쿼리로 삭제 기능 수정

+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,6 +1,7 @@
 package roomescape.service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import roomescape.dao.reservation.ReservationDao;
 import roomescape.domain.Reservation;
@@ -54,7 +55,10 @@ public class ReservationService {
                 .toList();
     }
 
-    public void delete(final Long id) {
+    public void delete(final long id) {
+        if (!reservationDao.existsById(id)) {
+            throw new NoSuchElementException("예약이 존재하지 않습니다.");
+        }
         reservationDao.delete(id);
     }
 }
